[SYSTEMDS-2832] Refactoring of old performance benchmarks
This commit extensively modify the performance benchmarks to use
the builtin functions. also added is arguments to execute the entire
benchmark within specific memory budgets.

DIA project WS2021/22

Closes #1481

In detail:

- Refactored old statistics benchmarks and changed them to use built-in functions.
- Improved logging management for benchmark outputs
- ALS conjugate gradient and direct solve benchmark with prediction.
- Added forced execution in specific folder.

+++ scripts/perftest/MatrixMult.sh
@@ -20,51 +20,56 @@
 #
 #-------------------------------------------------------------
 
-# Import MKL
-#if [ -d ~/intel ] && [ -d ~/intel/bin ] && [ -f ~/intel/bin/compilervars.sh ]; then
-#    . ~/intel/bin/compilervars.sh intel64
-#elif [ -d ~/intel ] && [ -d ~/intel/oneapi ] && [ -f ~/intel/oneapi/setvars.sh ]; then
-#	# For the new intel oneAPI
-#    . ~/intel/oneapi/setvars.sh intel64
-#else
-#    . /opt/intel/bin/compilervars.sh intel64
-#fi
+if [ "$(basename $PWD)" != "perftest" ];
+then
+  echo "Please execute scripts from directory 'perftest'"
+  exit 1;
+fi
 
-# Set properties
-#export LOG4JPROP='scripts/perftest/conf/log4j-off.properties'
-#export SYSDS_QUIET=1
-#export SYSTEMDS_ROOT=$(pwd)
-#export PATH=$SYSTEMDS_ROOT/bin:$PATH
+if ! command -v perf &> /dev/null
+then
+  echo "Perf stat not installed for matrix operation benchmarks, see README"
+  exit 0;
+fi
 
-
+CMD=$1
 
 # Logging output
-LogName='results/MM.log'
-mkdir -p 'results'
+LogName='logs/MM.log'
 rm -f $LogName
 
+tstart=$(date +%s.%N)
 # Baseline
 perf stat -d -d -d -r 5 \
-    systemds scripts/MM.dml \
+    ${CMD} scripts/MM.dml \
     -config conf/std.xml \
     -stats \
     -args 5000 5000 5000 1.0 1.0 3 \
     >>$LogName 2>&1
+ttrain=$(echo "$(date +%s.%N) - $tstart - .4" | bc)
+echo "Matrix mult 5000x5000 %*% 5000x5000 without mkl/openblas:" $ttrain >> results/times.txt
 
+
+tstart=$(date +%s.%N)
 # MKL
 perf stat -d -d -d -r 5 \
-    systemds scripts/MM.dml \
+    ${CMD} scripts/MM.dml \
     -config conf/mkl.xml \
     -stats \
     -args 5000 5000 5000 1.0 1.0 3 \
     >>$LogName 2>&1
+ttrain=$(echo "$(date +%s.%N) - $tstart - .4" | bc)
+echo "Matrix mult 5000x5000 %*% 5000x5000 with mkl:" $ttrain >> results/times.txt
 
+tstart=$(date +%s.%N)
 # Open Blas
 perf stat -d -d -d -r 5 \
-    systemds scripts/MM.dml \
+    ${CMD} scripts/MM.dml \
     -config conf/openblas.xml \
     -stats \
     -args 5000 5000 5000 1.0 1.0 3 \
     >>$LogName 2>&1
+ttrain=$(echo "$(date +%s.%N) - $tstart - .4" | bc)
+echo "Matrix mult 5000x5000 %*% 5000x5000 with openblas:" $ttrain >> results/times.txt
 
-cat $LogName | grep -E ' ba\+\* |Total elapsed time|-----------| instructions |  cycles | CPUs utilized ' | tee $LogName.log
+cat $LogName | grep -E ' ba\+\* |Total elapsed time|-----------| instructions |  cycles | CPUs utilized ' >> $LogName.log
